# docs/swagger.yaml
basePath: /
definitions:
...
  dto.TeamRequest:
    properties:
      description:
        type: string
      ispublic:
        type: boolean
      name:
        type: string
      teamtopic:
        $ref: '#/definitions/model.TopicOfInterest'
      userid:
        type: string
    type: object
...
  dto.UserToTeamRequest:
    properties:
      teamId:
        type: string
      userId:
        type: string
    required:
    - teamId
    - userId
    type: object
...
  entity.Team:
    properties:
      description:
        type: string
      id:
        type: string
      ispublic:
        type: boolean
      name:
        type: string
      teamtopic:
        $ref: '#/definitions/model.TopicOfInterest'
      users:
        items:
          type: string
        type: array
    type: object
...
  model.TopicOfInterest:
    enum:
    - Mathematics
    - Science
    - ComputerScience
    - Programming
    - Art
    - Music
    - Photography
    - Language
    type: string
    x-enum-varnames:
    - Mathematics
    - Science
    - ComputerScience
    - Programming
    - Art
    - Music
    - Photography
    - Language
info:
  contact: {}
  title: StudyWithMe API
  version: "1.0"
paths:
...
  /teams:
    get:
      description: Get teams - all teams, by name, or by prefix with limit
      parameters:
      - description: Filter by exact name
        in: query
        name: name
        type: string
      - description: Filter by name prefix
        in: query
        name: prefix
        type: string
      - description: Limit results (required with prefix)
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/entity.Team'
            type: array
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      security:
      - Bearer: []
      summary: Get teams with optional filtering
      tags:
      - teams
    post:
      consumes:
      - application/json
      description: Create a new team with the provided details
      parameters:
      - description: Team details
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.TeamRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/entity.Team'
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      security:
      - Bearer: []
      summary: Create a new team
      tags:
      - teams
  /teams/{id}:
    delete:
      description: Delete a team by providing team ID
      parameters:
      - description: Team ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Team deleted
          schema:
            additionalProperties: true
            type: object
        "400":
          description: 'Bad Request: Missing team ID'
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      security:
      - Bearer: []
      summary: Delete a team
      tags:
      - teams
    get:
      description: Get team details by ID
      parameters:
      - description: Team ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/entity.Team'
        "404":
          description: Team not found
          schema:
            additionalProperties: true
            type: object
      security:
      - Bearer: []
      summary: Get a team by ID
      tags:
      - teams
    put:
      consumes:
      - application/json
      description: Update team details by providing team ID and updated details
      parameters:
      - description: Team ID
        in: path
        name: id
        required: true
        type: string
      - description: Updated team details
        in: body
        name: team
        required: true
        schema:
          $ref: '#/definitions/entity.Team'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/entity.Team'
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      security:
      - Bearer: []
      summary: Update a team
      tags:
      - teams
  /teams/users:
    delete:
      description: Delete a user from a team by providing team ID
      parameters:
      - description: User ID and Team ID
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.UserToTeamRequest'
      produces:
      - application/json
      responses:
        "200":
          description: User deleted from team
          schema:
            additionalProperties: true
            type: object
        "400":
          description: 'Bad Request: Invalid request body or missing userId or teamId'
          schema:
            additionalProperties: true
            type: object
      security:
      - Bearer: []
      summary: Delete a user from a team
      tags:
      - teams
    put:
      consumes:
      - application/json
      description: Add a user to a team by providing user ID and team ID
      parameters:
      - description: User ID and Team ID
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.UserToTeamRequest'
      produces:
      - application/json
      responses:
        "200":
          description: User added to team
          schema:
            additionalProperties: true
            type: object
        "400":
          description: 'Bad Request: Invalid request body or missing userId or teamId'
          schema:
            additionalProperties: true
            type: object
      security:
      - Bearer: []
      summary: Add a user to a team
      tags:
      - teams
...
securityDefinitions:
  Bearer:
    description: 'Bearer JWT token. Example: "Bearer YOUR_TOKEN"'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"